+++ public/templates/show_playlist_row.inc.php
@@ -106,7 +106,13 @@ use Ampache\Module\Util\ZipHandlerInterface;
             echo Share::display_ui('playlist', $libitem->id, false);
         }
     }
-    if ($libitem->has_access()) { ?>
+    if ($libitem->has_access()) {
+        $search_id = $libitem->has_search();
+        if ($search_id > 0) { ?>
+            <a href="<?php echo AmpConfig::get('web_path'); ?>/playlist.php?action=refresh_playlist&type=playlist&id=<?php echo $libitem->id; ?>&search_id=<?php echo $search_id; ?>">
+                <?php echo Ui::get_icon('file_refresh', T_('Refresh from Smartlist')); ?>
+            </a>
+        <?php } ?>
     <a id="<?php echo 'edit_playlist_' . $libitem->id ?>" onclick="showEditDialog('playlist_row', '<?php echo $libitem->id ?>', '<?php echo 'edit_playlist_' . $libitem->id ?>', '<?php echo T_('Playlist Edit') ?>', 'playlist_row_')">
         <?php echo Ui::get_icon('edit', T_('Edit')); ?>
     </a>